server/requirements.txt
flask>=3.0.0
gitpython>=3.1.40
python-dotenv>=1.0.0
anthropic>=0.39.0
sentence-transformers>=2.2.0
